backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.0
openai==1.57.0
requests==2.32.0
openpyxl==3.1.5
pandas==2.2.3
pydantic==2.10.0
aiomysql==0.2.0
